資料管理集中 IDataStore
